LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.conic.negative.rotation.worker-expected.txt


FAIL Conic gradient with negative rotation assert_equals: Green channel of the pixel at (75, 40) expected 255 but got 254
